refactor: move word construction
